chat_engine.py
import os
import pandas as pd
from langchain.chat_models import ChatOpenAI
from langchain.agents import create_pandas_dataframe_agent
from langchain.agents.agent_types import AgentType
from dotenv import load_dotenv

load_dotenv()

def get_chat_response(df: pd.DataFrame, query: str) -> str:
    try:
        llm = ChatOpenAI(temperature=0, model="gpt-3.5-turbo", openai_api_key=os.getenv("OPENAI_API_KEY"))
        agent = create_pandas_dataframe_agent(
            llm,
            df,
            agent_type=AgentType.OPENAI_FUNCTIONS,
            verbose=False
        )
        response = agent.run(query)
        return response
    except Exception as e:
        return str(e)