requirements.txt
streamlit
pandas
langchain
openai
python-dotenv